+++ src/App.js
@@ -39,71 +39,50 @@ class App extends Component {
     const { name } = this.props.userData;
         // console.log(name, 'me mayne');
     return (
-      <Router class="ui five item menu">
-        <div>
-          <Sidebar as={Menu} animation="push" width="thin" visible icon="labeled" vertical inverted>
+      <div>
+        <Sidebar as={Menu} animation="push" width="thin" visible icon="labeled" vertical inverted>
 
-            <Menu.Item name="home">
-              <Link to="/Home">
-                <Icon name="home" size="medium" />
-                <h3>
+          <Menu.Item name="home">
+            <Link to="/Home">
+              <Icon name="home" size="medium" />
+              <h3>
                               Home
                             </h3>
-              </Link>
-            </Menu.Item>
-            <Menu.Item name="cohort">
-              <Link to="/Cohort">
-                <Icon name="users" size="medium" />
-                <h3>
+            </Link>
+          </Menu.Item>
+          <Menu.Item name="cohort">
+            <Link to="/Cohort">
+              <Icon name="users" size="medium" />
+              <h3>
                               Cohort
                             </h3>
-              </Link>
-            </Menu.Item>
-            <Menu.Item name="profile">
-              <Link to="/Profile">
-                <Icon name="user" size="medium" />
-                <h3>
+            </Link>
+          </Menu.Item>
+          <Menu.Item name="profile">
+            <Link to="/Profile">
+              <Icon name="user" size="medium" />
+              <h3>
                                   Profile
                                 </h3>
-              </Link>
-            </Menu.Item>
+            </Link>
+          </Menu.Item>
 
-            <Menu.Item name="login">
-              <Link to="/">
-                <Icon name="sign in" size="medium" />
-                <h3>
+          <Menu.Item name="login">
+            <Link to="/">
+              <Icon name="sign in" size="medium" />
+              <h3>
                               Login
                             </h3>
-              </Link>
-            </Menu.Item>
-          </Sidebar>
-          <Route
-            path="/Home" render={() => (!name
-                        ? (<Redirect to="/Login" />)
-                        : (<Home />))}
-          />
-          <Route path="/Login" component={GoogleAuthentication} />
-          <Route
-            exact path="/" render={() => (name
-                        ? (<Redirect to="/Home" />)
-                        : (<Redirect to="/Login" />))}
-          />
-          <Route
-            path="/Cohort" render={() => (!name
-                        ? (<Redirect to="/Login" />)
-                        : (<Cohort />))}
-          />
-          <Route
-            path="/Profile" render={() => (!name
-                        ? (<Redirect to="/Login" />)
-                        : (<Profile />))}
-          />
-        </div>
-      </Router>
+            </Link>
+          </Menu.Item>
+        </Sidebar>
+
+
+      </div>
+
     );
   }
 
 }
 
-const mapStateToProps = ({ userData }) => ({ userData });
-export default connect(mapStateToProps)(App);
+export default connect(({ userData }) => ({ userData }))(App);
