+++ src/components/HOC/Require_Authentication.js
@@ -1,10 +1,14 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
+import Login from '../../Views/Login'
 
 export default function (ComposedComponent) {
  class Authentication extends Component {
+  constructor(props) {
+   super(props)
+  }
   static contextTypes = {
-   router: React.PropTypes.object
+   router: React.PropTypes.object,
   }
   componentWillMount = () => {
     if(!this.props.authentication) {
